+++ frameworks/react/components/display/calendar-event/CalendarEvent.d.ts
@@ -13,6 +13,8 @@ export interface CalendarEventProps {
 
   colorId?: CatSlot;
 
+  interactive?: boolean;
+
   disabled?: boolean;
 
   actionsEnabled?: boolean;